backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
PyJWT==2.8.0
pandas==2.1.3
numpy==1.26.2
pyarrow==14.0.1
scikit-learn==1.3.2
xgboost==2.0.2
scipy==1.11.4
httpx==0.25.2
python-dotenv==1.0.0
